Avoid empty readme error, and dont show top docstring (to keep focus on the API)
By removing the toplevel docstring (which is where the user should write
general intro, install, usage, changes, etc) the focus of the html docs
site is on API only. This is necessary because currently output html
does not provide header links to the intro/tutorial sections.

Ideally in future, html docs should have two pages: the docstring
intro/tutorial section with a link and sublinks, and an API section with
a link and sublinks. Maybe possible by using the pdoc "external links"
object?